requirements.txt
ml-collections==0.1.0
absl-py==0.10.0
torch>=2.0.0
torchvision>=0.15.0
ninja
pillow
scipy
numpy